+++ CHANGELOG.md
@@ -77,6 +77,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Checks for sheet name in L1 input validation and generation. [#285](https://github.com/ncsa/ameriflux-pipeline/issues/285)
 - Regex matching for met tower variable names. [#289](https://github.com/ncsa/ameriflux-pipeline/issues/289), [#299](https://github.com/ncsa/ameriflux-pipeline/issues/299)
 - Documentation files. [#297](https://github.com/ncsa/ameriflux-pipeline/issues/297)
+- Documentation on metprocessor. [#308](https://github.com/ncsa/ameriflux-pipeline/issues/308), [#309](https://github.com/ncsa/ameriflux-pipeline/issues/309)
 - Documentation on enveditor. [#298](https://github.com/ncsa/ameriflux-pipeline/issues/298)
 - Key file for variable renaming in met merger. [#301](https://github.com/ncsa/ameriflux-pipeline/issues/301)
 - Documentation on prepyfluxpro module. [#311](https://github.com/ncsa/ameriflux-pipeline/issues/311)
@@ -133,3 +134,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Ameriflux sitename updated. [#302](https://github.com/ncsa/ameriflux-pipeline/issues/302)
 - Deleted unneccessary test modules. [#300](https://github.com/ncsa/ameriflux-pipeline/issues/300)
 - Metmerger modules renamed. [#312](https://github.com/ncsa/ameriflux-pipeline/issues/312)
+- Added key file process in metprocessor. [#316](https://github.com/ncsa/ameriflux-pipeline/issues/316)
+- Updated the texts in enveditor. [#324](https://github.com/ncsa/ameriflux-pipeline/issues/324)